Refactor `all_modules` to be included in utils (fixes previous docs failing) (#1061)
