+++ docs/_scripts/graph_gen.py
@@ -2,7 +2,8 @@ import math
 import os
 import subprocess
 from collections import defaultdict
-from test.all_modules import all_environments
+
+from pettingzoo.utils.all_modules import all_environments
 
 
 def generate_cycle_words(agents, is_classic):
