+++ docs/_scripts/text_capture.py
@@ -1,8 +1,9 @@
 import json
 import random
-from test.all_modules import all_environments
 
 import numpy as np
+
+from pettingzoo.utils.all_modules import all_environments
 
 
 def shrink_lines(data, max_height, max_width):
