+++ test/unwrapped_test.py
@@ -1,9 +1,8 @@
-from test.all_modules import all_environments
-
 import pytest
 from gymnasium import spaces
 
 from pettingzoo.utils import conversions, wrappers
+from pettingzoo.utils.all_modules import all_environments
 
 
 def box_action(env, agents):
